somechanges 2016/10/14

+++ delete.php
@@ -30,12 +30,12 @@
             }
 
 
-        }else{
+        }elseif($_POST['table'] == 'companies'){
 
             $compID = $_POST['id'];
             $table = $_POST['table'];
             $query = "DELETE FROM ".$table." WHERE id='".$compID."'";
-            echo ($query); #query hewlej uzsen console deer
+            #echo ($query); #query hewlej uzsen console deer
 
             if($conn->query($query)) {
                 $query = "DELETE FROM employee WHERE companyID='".$compID."'";
@@ -49,6 +49,27 @@
             }
 
             echo json_encode($content);
+
+        }elseif($_POST['table'] == 'material') {
+
+            $clientId = $_POST['id'];
+            $query = "DELETE FROM client WHERE id='".$clientId."'";
+            #echo ($query); #query hewlej uzsen console deer
+
+            if($conn->query($query)) {
+                $query = "DELETE FROM material WHERE clientID ='".$clientId."'";
+                $conn->query($query);
+                #echo ($query); #query hewlej uzsen console deer
+                $content->success = true;
+                $content->response = "Падаан устгагдалаа";
+            }else{
+                $content->success = false;
+                $content->response = "";
+            }
+
+            echo json_encode($content);
+            
+
         }
     #table request end
     }
